requirements.txt
aiogram
google-genai
pandas
pillow
schedule
openpyxl